+++ HISTORY.md
@@ -7,6 +7,7 @@
 - Added ability to combine classification inference results from different architectures
 - Added ability to save `initial` and `latest` models in addition to `best`
 - Added ability to specify testing data csv in main cli
+- Normalized surface dice has been added
 
 ## 0.0.15
 - Updated `setup.py` for `python>=3.8`